+++ packages/server/src/index.ts
@@ -1,1 +1,11 @@
-export { startServer } from "./server/start-server";
+import { config } from "dotenv";
+config({ path: ".env.gateweaver" });
+import { logger } from "@gateweaver/utils";
+import { startServer } from "./start-server";
+
+try {
+  startServer();
+} catch (error) {
+  logger.error(error);
+  process.exit(1);
+}