+++ packages/server/tsup.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/index.ts", "src/start-server.ts"],
+  format: ["cjs", "esm"],
+  dts: true,
+});