+++ packages/policies/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @gateweaver/policies
 
+## 0.0.4
+
+### Patch Changes
+
+- Publish packages
+- Updated dependencies
+  - @gateweaver/utils@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Update server dist
+- Updated dependencies
+  - @gateweaver/utils@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Small patch
+- Updated dependencies
+  - @gateweaver/utils@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes