+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gateweaver/utils
 
+## 0.0.4
+
+### Patch Changes
+
+- Publish packages
+
+## 0.0.3
+
+### Patch Changes
+
+- Update server dist
+
+## 0.0.2
+
+### Patch Changes
+
+- Small patch
+
 ## 0.0.1
 
 ### Patch Changes